refactor: [EL-4435] extract conversation starters into useConversationStarters hook
- Move conversationStarters and editorConversationStarters state to custom hook
- Derive displayedConversationStarters logic in dedicated hook
- Reduce NewChat component complexity (reduced ~40 lines)
