+++ src/[fsd]/features/chat/lib/hooks/index.js
@@ -1,4 +1,5 @@
 export { useConversationNavigation } from './useConversationNavigation.hooks';
+export { useConversationStarters } from './useConversationStarters.hooks';
 export { useEditConversation } from './useEditConversation.hooks';
 export { useInternalToolsConfig } from './useInternalToolsConfig.hooks';
 export { useAttachmentToolChange } from './useAttachmentToolChange.hooks';
